// apps/capsulelabs-api/src/star-time-capsule/enums/astronomical-event.enum.ts
export enum AstronomicalEventType {
  FULL_MOON = "full_moon",
  NEW_MOON = "new_moon",
  SUMMER_SOLSTICE = "summer_solstice",
  WINTER_SOLSTICE = "winter_solstice",
  SPRING_EQUINOX = "spring_equinox",
  AUTUMN_EQUINOX = "autumn_equinox",
  LUNAR_ECLIPSE = "lunar_eclipse",
  SOLAR_ECLIPSE = "solar_eclipse",
  METEOR_SHOWER = "meteor_shower",
  PLANETARY_ALIGNMENT = "planetary_alignment",
}